+++ src/App.css
@@ -38,120 +38,3 @@
     transform: rotate(360deg);
   }
 }
-
-
-.footer{
-  display: flex;
-  flex-direction: baseline;
-  justify-content:space-around;
-}
-
-
-
-/* ________________---FOOTER---________________*/
-
-.Footer{
-  width: 100%;
-  background-color: #061726;
-
-}
-.li{
-  list-style: none;
-  flex-direction: column;
-  text-align: left;
-}
-.img{
-  width: 40px;
-  height: 40px;
-}
-
-
-
-/* ________________---GRUPO 0---________________*/
-
-.Footer .grupo-0{
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(5,1fr);
-  grid-gap: 10px;
-  padding: 10px 0px;
-  background-color: rgb(170, 167, 167);
-  text-align: center;
-}
-/* ________________---GRUPO 1---________________*/
-
-.Footer .grupo-1{
-  width: 100%;
-  max-width: 1200px;
-  margin: auto;
-  display: grid;
-  grid-template-columns: repeat(4,1fr);
-  grid-gap: 50px;
-  padding: 30px 0px;
-}
-
-.Footer .grupo-1 .imgLogoFooter{
-  width: 250px;
-}
-
-/* ________________---BOX GRUPO 1---________________*/
-
-.Footer .grupo-1 .box h3{
-  color: #fff;
-  margin-bottom: 25px;
-  font-size: 20px;
-}
-
-
-
-.contactImg{
-  width: 25px;
-  height: 25px;
-  float: left;
-  margin: 1px;
-}
-
-.logoImg{
-  width: 45px;
-  height: 45px;
-}
-
-
-.logoImg{
-  width: 150px;
-  height: 150px;
-}
-
-
-.Footer .grupo-1 .box p{
-  color: rgb(174, 178, 181);
-  margin-bottom: 10px;
-}
-
-.Footer .grupo-1 .box p:hover{
-  color: white;
-}
-
-/* ________________---RED SOCIAL---________________*/
-
-.Footer .grupo-1 .redSocial img{
-  display: inline-block;
-  text-decoration: none;
-  width: 60px;
-  height: 60px;
-  margin-right: 15px;
-}
-
-.Footer .grupo-1 .redSocial img:hover{
-  filter: opacity(0.3) drop-shadow(0 0 0 rgb(215, 2, 201)); 
-}
-
-/* ________________---GRUPO 2---________________*/
-
-.Footer .grupo-2{
-  background-color: #052a4a ;
-  padding: 15px 10px;
-  text-align: center;
-  color: white;
-}
-
