+++ src/Componentes/Footer/Footer.css
@@ -1,0 +1,74 @@
+.Footer{
+  width: 100%;
+  background-color: #7b078bce;
+}
+
+  /* ________________---GRUPO 1---________________*/
+
+  .Footer .grupo-1{
+    width: 100%;
+    max-width: 1600px;
+    margin: auto;
+    display: grid;
+    grid-template-columns: repeat(5,1fr);
+    grid-gap: 40px;
+    padding: 30px 0px;
+  }
+  .Footer .grupo-1 .imgLogoFooter{
+    width: 250px;
+    height: 250px;
+  }
+  /* ________________---BOX GRUPO 1---________________*/
+
+  .Footer .grupo-1 .box h3{
+    color: rgb(0, 0, 0);
+    margin-bottom: 30px;
+    font-size: 20px;
+  }
+
+  .Footer .grupo-1 .box p{
+    color: rgba(178, 178, 178, 0.717);
+    margin-bottom: 10px;
+    font-size: 23px;
+  }
+
+  .Footer .grupo-1 .box .resaltado:hover{
+    color: rgb(229, 228, 228);
+  }
+
+  /* ________________---RED SOCIAL---________________*/
+
+
+  .Footer .grupo-1 .redSocial img{
+    display: flex;
+    text-decoration: none;
+    width: 50px;
+    height: 50px;
+    margin-right: 15px;
+    align-items:flex-start;
+    padding: 3px;
+  }
+
+  .Footer .grupo-1 .redSocial img:hover{
+    filter: opacity(0.3) drop-shadow(0 0 0 #70027ece);
+  }
+
+  /* ________________---GRUPO 2---________________*/
+
+  .Footer .grupo-2 {
+    background-color: #70027ece ;
+    padding: 15px 10px;
+    text-align: center;
+    color: white;
+  }
+
+  /* ________________---RESPONSIVE---________________*/
+
+  @media screen and (max-width:1000px){
+    .Footer .grupo-1{
+        width: 90%;
+        grid-template-columns: repeat(1, 1fr);
+        grid-gap:30px;
+        padding: 35px 0px;
+    }
+}
